src/main/java/co/com/softka/softkau/tengohambrerestaurantbar/usecase/factura/CrearFacturaUseCase.java
package co.com.softka.softkau.tengohambrerestaurantbar.usecase.factura;

import co.com.sofka.business.generic.UseCase;
import co.com.sofka.business.support.RequestCommand;
import co.com.sofka.business.support.ResponseEvents;
import co.com.softka.softkau.tengohambrerestaurantbar.domain.factura.Factura;
import co.com.softka.softkau.tengohambrerestaurantbar.domain.factura.commands.CrearFactura;

public class CrearFacturaUseCase extends UseCase<RequestCommand<CrearFactura>, ResponseEvents> {

    @Override
    public void executeUseCase(RequestCommand<CrearFactura> crearFacturaRequestCommand) {
        var command = crearFacturaRequestCommand.getCommand();
        var factura = new Factura(command.getFacturaId(), command.getFecha());
        emit().onResponse(new ResponseEvents((factura.getUncommittedChanges())));
    }
}
